Using Material UI, TS, Leaflet
Starting to use Material UI components more on the site as their hooks css API and pre-made components are very useful. Adding Typescript support for now but might remove later. Leaflet-react was also added as a dependency as the plan is to add an interactive map using the JHU coordinates for providence and country data.