Adding environment variables setup
Some API tokes are needed but it is bad practice to leave them open. Right now the solution involves having a local json file with the API tokens and using a python script to load the files onto the next.config.js file and unload them back to the default once the server has ended.