Chaning of JHU Github Repo
The main information repository where I gather the information for the backend db recently changed the way that data is uploaded. The previous three repositories that where used to gather confirmed, dead, and recovered cases where deprecated in favor of a new system with slight changes in what cruise ships get included in the statistics. Regardless the changes required where minimal and now the db lacks a direct way to estimate the recovered cases without looking at other sources. an old file of cases is included just in case.